crud terminado y bd agregada

+++ WebApiVentas.Core/Models/LugarEntrega.cs
@@ -14,7 +14,7 @@ namespace WebApiVentas.Core.Models
 
         public int LugarEntregaId { get; set; }
         public int DistritoId { get; set; }
-        public decimal PrecioEnvio { get; set; }
+        public double PrecioEnvio { get; set; }
 
         public virtual Distrito Distrito { get; set; }
         public virtual ICollection<Pedido> Pedidos { get; set; }
